updated column money to bigInt

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -22,7 +22,7 @@ return new class extends Migration
             $table->rememberToken();
             $table->integer('tickets_created')->default(0);
             $table->integer('roll_events_participated')->default(0);
-            $table->integer('money')->default(0);
+            $table->bigInteger('money')->default(0);
             $table->timestamps();
         });
     }
